Refactor and expand test suite structure
Moved and split tests into a dedicated 'tests' package, separating database and query tests into 'test_db.py' and 'test_query.py'. Added shared fixtures in 'conftest.py', constants in 'constants.py', and utility functions in 'utils.py' to improve test maintainability and reusability.
