+++ backend/tests/constants.py
@@ -1,0 +1,7 @@
+from router.file_api import BASE_DIR
+
+PATH_HUSKY_IMAGE = "husky_1.jpg"
+PATH_HUSKY_IMAGE_2 = "folder/husky_2.jpg"
+PATH_ROBOT_IMAGE = "robot_1.jpg"
+PATH_ROBOT_IMAGE_2 = "folder/robot_2.jpg"
+PATH_FLOWER_IMAGE = "flower.jpg"
